Add Cortex M3 NVIC
