+++ qiling/hw/intc/__init__.py
@@ -3,6 +3,7 @@
 # Cross Platform and Multi Architecture Advanced Binary Emulation Framework
 #
 
+from .cm3_nvic import CortexM3Nvic
 from .cm4_nvic import CortexM4Nvic
 from .stm32f4xx_exti import STM32F4xxExti
 from .gd32vf1xx_eclic import GD32VF1xxEclic
