+++ qiling/hw/intc/cm3_nvic.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# 
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+#
+
+from qiling.hw.intc.cm_nvic import CortexMNvic
+
+class CortexM3Nvic(CortexMNvic):
+    pass
